backend/requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic[email]==2.9.0
pydantic-settings==2.5.0

# Database
sqlalchemy==2.0.35
alembic==1.13.0
psycopg2-binary==2.9.9   # PostgreSQL driver for production


# Auth
bcrypt==4.2.0
PyJWT==2.9.0

# Twilio (voice + SMS)
twilio==9.3.0

# Email
sendgrid==6.11.0

# AI/LLM (optional — needed for voice/chat agent)
anthropic==0.39.0
openai==1.50.0

# Scheduling
apscheduler==3.10.4

# Dev/testing
httpx==0.27.0
pytest==8.3.0
pytest-asyncio==0.24.0
